Move earlystopper to its own file
