+++ plugins/hydra_ax_sweeper/hydra_plugins/hydra_ax_sweeper/ax_sweeper.py
@@ -16,6 +16,8 @@ from hydra.plugins.search_path_plugin import SearchPathPlugin
 from hydra.plugins.sweeper import Sweeper
 from hydra.types import TaskFunction
 
+from .earlystopper import EarlyStopper
+
 log = logging.getLogger(__name__)
 
 
@@ -80,63 +82,6 @@ def get_one_batch_of_trials(
             )
         )
     return batch_of_trials
-
-
-class EarlyStopper:
-    """Class to implement the early stopping mechanism.
-    The optimisation process is stopped when the performance does not
-    improve for a threshold number of consecutive epochs. The performance
-    is considered to have improved when the change is more than a given
-    threshold (epsilon)."""
-
-    def __init__(
-        self, max_epochs_without_improvement: int, epsilon: float, minimize: bool
-    ):
-        self.max_epochs_without_improvement = max_epochs_without_improvement
-        self.epsilon = epsilon
-        self.minimize = minimize
-        self.current_best_value: Optional[float] = None
-        self.current_epochs_without_improvement = 0
-
-    def should_stop(
-        self, potential_best_value: float, best_parameters: ParameterType
-    ) -> bool:
-        """Check if the optimisation process should be stopped."""
-        is_improving = True
-        if self.current_best_value is not None:
-            if self.minimize:
-                is_improving = (
-                    potential_best_value + self.epsilon < self.current_best_value
-                )
-            else:
-                is_improving = (
-                    potential_best_value - self.epsilon > self.current_best_value
-                )
-
-        if is_improving:
-            self.current_epochs_without_improvement = 0
-            self.current_best_value = potential_best_value
-            log.info(
-                "New best value: {}, best parameters: {}".format(
-                    potential_best_value, best_parameters
-                )
-            )
-
-            return False
-        else:
-            self.current_epochs_without_improvement += 1
-
-        if (
-            self.current_epochs_without_improvement
-            >= self.max_epochs_without_improvement
-        ):
-            log.info(
-                "Early stopping, best known value {} did not improve for {} epochs".format(
-                    self.current_best_value, self.current_epochs_without_improvement
-                )
-            )
-            return True
-        return False
 
 
 class AxSweeper(Sweeper):
